SSR specs and optimizing resource bundling
