+++ packages/cli/test/cases/develop.ssr/develop.ssr.spec.js
@@ -232,7 +232,7 @@ describe('Develop Greenwood With: ', function() {
         const counterScript = Array.from(dom.window.document.querySelectorAll('head > script[src]'))
           .filter((tag) => tag.getAttribute('src').indexOf(`${componentName}.js`) >= 0);
 
-        expect(artistsPageGraphData.imports[0]).to.equal(`/components/${componentName}.js`);
+        expect(artistsPageGraphData.imports[0].src).to.equal(`/components/${componentName}.js`);
         expect(counterScript.length).to.equal(1);
       });
     });
